// enums/document-type.ts
export enum DocumentType {
  UPLOADED = 'UPLOADED',
  FINANCIAL_STATEMENT = 'FINANCIAL_STATEMENT',
  FINANCIAL_ANALYSIS = 'FINANCIAL_ANALYSIS',
  BANK_STATEMENT = 'BANK_STATEMENT',
  BANK_ANALYSIS = 'BANK_ANALYSIS',
  CONSOLIDATED_ANALYZE = 'CONSOLIDATED_ANALYZE',
  COMBINED_ANALYZE = 'COMBINED_ANALYZE',
  FRAUD_DETECTION = 'FRAUD_DETECTION',
}

export enum tab {
  UPLOADED = 'UPLOADED',
  FINANCIAL_STATEMENT = 'FINANCIAL_STATEMENT',
  BANK_STATEMENT = 'BANK_STATEMENT',
  COMBINED = 'COMBINED',
}